refactor: ImageSize 분리
