+++ src/test/java/nextstep/courses/domain/cource/ImageTest.java
@@ -23,14 +23,4 @@ class ImageTest {
     public void fileSize_check(){
         Assertions.assertThatThrownBy(() -> Image.of("jpg", 300L, 200L, 1025)).isInstanceOf(ImageFileSizeExceededException.class);
     }
-
-    @Test
-    @DisplayName("이미지 크기는 width 300, height 200 이상이어야 하며, 비율은 3:2 이여야 한다.")
-    public void check_imageSize_and_ratio(){
-        Assertions.assertThatThrownBy(() -> Image.of("jpg", 299L, 200L, 1024)).isInstanceOf(ImageSizeBelowMinException.class);
-
-        Assertions.assertThatThrownBy(() -> Image.of("jpg", 299L, 199L, 1024)).isInstanceOf(ImageSizeBelowMinException.class);
-
-        Assertions.assertThatThrownBy(() -> Image.of("jpg", 300L, 201L, 1024)).isInstanceOf(UnsupportedImageRatioException.class);
-    }
 }
